# swagger-service/public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/admin/manual_overrides": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List manual overrides",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Surface",
                        "name": "surface",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (active,cancelled,expired)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by action (boost,suppress)",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include expired overrides",
                        "name": "include_expired",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.ManualOverrideResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a manual boost/suppression override",
                "parameters": [
                    {
                        "description": "Manual override payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ManualOverrideRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.ManualOverrideResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    }
                }
            }
        },
        "/v1/admin/manual_overrides/{override_id}/cancel": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Cancel a manual override",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Override ID",
                        "name": "override_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional cancellation metadata",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.ManualOverrideCancelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ManualOverrideResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.APIError": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "types.ManualOverrideCancelRequest": {
            "type": "object",
            "properties": {
                "cancelled_by": {
                    "type": "string"
                }
            }
        },
        "types.ManualOverrideRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "boost_value": {
                    "type": "number"
                },
                "created_by": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "item_id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "surface": {
                    "type": "string"
                }
            }
        },
        "types.ManualOverrideResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "boost_value": {
                    "type": "number"
                },
                "cancelled_at": {
                    "type": "string"
                },
                "cancelled_by": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "item_id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "override_id": {
                    "type": "string"
                },
                "rule_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "surface": {
                    "type": "string"
                }
            }
        },
...
    }
}